ghost-to-wordpress/requirements.txt
python-wordpress-xmlrpc>=2.3
